Move wait_for() from app_test/main to a hpp file
